# PERF: optimize coordinate access for scalar Point