Illustration of what it would entail to optimize specifically the `Point.x` et al attributes for scalar points.

```
In [1]: import shapely

In [2]: p = shapely.Point(12, 34)

In [3]: %timeit p.x
3.15 µs ± 3.46 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)    # <-- main
236 ns ± 0.164 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)  # <-- PR
```

The reason this is so much faster is mostly from avoiding the ufunc machinery overhead.

While this is relatively little code for just those `x`/`y`/`z` attributes, the question is of course then "could we also speed up this other attribute .. ?", and we definitely don't want to end up reimplementing all ufuncs for scalar geometries.