@theroggy 2024-07-25
> While this is relatively little code for just those `x`/`y`/`z` attributes, the question is of course then "could we also speed up this other attribute .. ?", and we definitely don't want to end up reimplementing all ufuncs for scalar geometries.

I definitely understand this...

Then again, there are cases where a vectorized approach is not feasable, e.g. when implementing more complicated algorithms using many basic spatial operations on a single geometry. If a shortcut that bypasses the ufunc machinery if the input is a single element would give a big performance difference...