@theroggy 2025-10-19
Algorithms like [centerline](https://github.com/pygeoops/pygeoops/blob/main/pygeoops/_centerline.py) are not really feasible to implement without many calls on scalar functions I think. Could be implemented in cython of course, but...

In an ideal world the optimized scalar path would be implemented in c I suppose so it can be tackled centrally, at least for all functions that don't need custom code in ufunc.c? But I don't know that code at all, so possibly/probably this is not possible?